backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
sqlalchemy[asyncio]==2.0.23
pymysql==1.1.0
aiomysql==0.2.0
alembic==1.13.0

# Redis & Caching
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation
email-validator==2.1.0

# Logging
python-json-logger==2.0.7

# Task Queue (Celery)
celery==5.3.4
flower==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Performance Monitoring
prometheus-client==0.19.0

# CORS
fastapi-cors==0.0.6
